validação de exclusão de veiculos não alugados

+++ src/services/customer/DeleteCustomerService.ts
@@ -4,6 +4,12 @@ import { customerRepository } from "../../repositories/CustomerRepository";
 
 class DeleteCustomerService {
     execute(id: string): void {
+        const customer = customerRepository.getById(id);
+
+        if (customer.hasRent) {
+            throw new AppError("It is not possible to delete customers with rents", StatusCodes.BAD_REQUEST);
+        }
+
         const deletedCustomer = customerRepository.delete(id);
 
         if (!deletedCustomer) {
